feat(battery): created a single battery for youtube
